project/requirements.txt
# Framework Web
FastAPI==0.115.11
uvicorn==0.24.0
Flask==3.0.3
gunicorn==20.1.0

# Base de Données
SQLAlchemy==2.0.23
psycopg2==2.9.10

# Manipulation des Données
pandas==2.1.1
numpy==1.26.0

# Visualisation
matplotlib==3.10.1
plotly==5.17.0
seaborn==0.13.2
mpld3==0.5.10

# Traitement Vidéo
opencv-python==4.8.1.78
ffmpeg-python==0.2.0
pymediainfo==7.0.1

# Traitement JSON
jsonschema==4.19.0
python-multipart==0.0.20
PyYAML==6.0.2

# Développement et Outils
python-dotenv==1.0.0
pytest==8.3.5
pytest-asyncio==0.26.0
iniconfig==2.1.0
pluggy==1.5.0
tenacity==9.0.0
tqdm==4.67.1
rich==13.9.4
rich-toolkit==0.13.2
tzdata==2025.1

# Utilitaires Divers
aiohttp==3.11.14
attrs==25.1.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
email-validator==2.2.0
fonttools==4.55.7
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib-metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
multidict==6.2.0
nest-asyncio==1.6.0
packaging==24.2
pillow==11.1.0
pip==25.0
propcache==0.3.1
pydantic==2.10.6
pydantic-core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
referencing==0.36.2
requests==2.32.3
retrying==1.3.4
rpds-py==0.22.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
typing-extensions==4.12.2
urllib3==2.3.0
watchfiles==1.0.4
websockets==15.0.1
Werkzeug==3.0.6
wheel==0.44.0
yarl==1.18.3
zipp==3.21.0

# Dataset du projet
lsfb-dataset==2.0.1
